# Conflicts:
#	maps/10_player/TERRETORIALE HERAUSFORDERUNG.map
#	maps/14_player/7VS7 GAME.map
#	maps/2_player/(2) - Coastside.map
#	maps/2_player/(2) - River of War.map
#	maps/2_player/60-ZWEITER KRIEG.map
#	maps/2_player/ALARA RANGE.map
#	maps/2_player/ASPHALT ROAD.map
#	maps/2_player/ATLAS.map
#	maps/2_player/BEAKER RIVER.map
#	maps/2_player/COG ISLE.map
#	maps/2_player/CROSSFIRE.map
#	maps/2_player/Coastside.map
#	maps/2_player/EON SPRINGS.map
#	maps/2_player/GO ISLANDS.map
#	maps/2_player/GREEN MOON.map
#	maps/2_player/HOURGLASS ISLE.map
#	maps/2_player/LITTLE ISLAND.map
#	maps/2_player/MELON LAND.map
#	maps/2_player/NORTH STAR.map
#	maps/2_player/ORBITAL RING.map
#	maps/2_player/PORTAL BRIDGE.map
#	maps/2_player/RAINY HAVEN.map
#	maps/2_player/RIDGE ISLAND.map
#	maps/2_player/RUBY KEYS.map
#	maps/2_player/River of War.map
#	maps/2_player/S LAND.map
#	maps/2_player/SABRE RANGE.map
#	maps/2_player/SHANGRI-LA.map
#	maps/2_player/STAR ISLANDS.map
#	maps/2_player/SUN CANAL.map
#	maps/2_player/SWITCHBACK.map
#	maps/2_player/TRIANGLES.map
#	maps/2_player/VOLCANO ISLE.map
#	maps/2_player/ZERO WOOD.map
#	maps/5_player/AUGE DES HURRIKANS.map
#	maps/6_player/DELUXE DOLLAR ISLAND.map

+++ ingamescriptsupport/events/scripteventchangecobar.cpp
@@ -5,8 +5,6 @@
 
 #include "resource_management/fontmanager.h"
 
-#include "coreengine/mainapp.h"
-
 #include "objects/base/spinbox.h"
 #include "objects/base/label.h"
 
@@ -130,7 +128,7 @@ void ScriptEventChangeCOBar::showEditEvent(spScriptEditor pScriptEditor)
     connect(spinBox.get(), &SpinBox::sigValueChanged, this,
             [this](qreal value)
     {
-        setStars(static_cast<qint32>(value));
+        setStars(value);
     });
     pBox->addItem(spinBox);
